backend/requirements.txt
# Backend Python Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Alternative: FastAPI
# fastapi==0.109.0
# uvicorn==0.27.0

# AI/ML Libraries - Google Gemini
google-generativeai==0.3.2

# Hugging Face - Image Generation
huggingface-hub>=1.1.5
Pillow==10.2.0

# Database (choose based on your needs)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.1          # MongoDB

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3

# Optional: For advanced agent features
# chromadb==0.4.22        # Vector database
# numpy==1.26.3
# pandas==2.1.4
